+++ python-lib/language_dict.py
@@ -218,9 +218,7 @@ SUPPORTED_LANGUAGES_SPACY = {
     "id": "Indonesian",
     "is": "Icelandic",
     "it": "Italian",
-    "ja": "Japanese",
     "kn": "Kannada",
-    "ko": "Korean",
     "lb": "Luxembourgish",
     "lt": "Lithuanian",
     "lv": "Latvian",
